Something else (#2)
* fix(users): made user selection explicit

used text()="name" instead of contains

* fix(new user validation): user location now compatible with
new website layout

Finds user based on the name in the td instead of looking at the
id of the button.

* feat(convert image): check if error messages are cleared

Check if source and output error messages are cleared after
multiple erroneous inputs

* feat(logout): Add a logout after each test and some cleaning

after and afterEach logouts depending on whether there is a before
or beforeEach. Also some minor cleanup of unused lines

* feat(users): made xpath less likely to fail

xpath was overinclusive and could create misleading test results
tightened it up so it should only respond to proper conditions

* refactor(user validation): shortened code with create user function
from common

* fix(common): various fixes to common

Stopped checking if user creation succeeded in createNewUser,
left it up to tests call it. Cleaned up some xpath and added
a few waits to login and sidebar.

* fix(convert image): typo in xpath

* style(various): cleanup of empty space and unused code

* fix(users): typo fix

* test(realms): new tests for creation, deletion and overwriting
new realms.

* fix(users): fixed code to reflect changes made on the site as
well as using a new way to check that button onclick is not
null

* fix(new_users): longer wait time

* feat(common): waitAndClick for clicking on submit buttons
or buttons with an onclick value

Additionally, fixed waitUntil that was only calling the inner
function once

* fix(various): fixed graph navigation errors and added some
whitespace to make tests more readable

* test(pools): check if a template exists, test creation and deletion
if it does, otherwise check that creation is not allowed

* test(templates): basic ui testing for template creatyion
and addition

* test(template validations): check for blank and invalid inputs
in various fields for template creation and addition

* test(administration): basic ui testing for administration, runs
through all the buttons

* test(firmware): invalid submission test

* test(network interfaces): basic ui testing, cancel and submit for
hardcode and bonding

* chore(realms): added to do list and reformated tabs

* refactor(users): moved users.test.js into users folder and
renamed it basic

* refactor(new user validation): code responds with an expect instead
of a wait for failure and empty clean up test to be completed at
a later date

* feat(common): added branching statements for click side bar tab
and login to test whether it is necessary to do them

* chore(misc): added to do lists and minor edits

* feat(login): commented out login tests as they are duplicative
and unnecessary as well as redundant

* chore(misc): removed tabs from code and replaced them with two spaces

* chore(users and templates): restructured folder organization and
deleted duplicative files

* chore(template validation): more responsive before and added a
reminder test for cleanup

* fix(guest pools): testing without making sure there are available
templates, commenting out for now will be included in later
tests

* fix(guest pools): commented out file due to it being reliant on
templates

will reincorporate into integrated test